+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 All notable changes to CAPSEM will be documented in this file.
 
-## [1.1.1] - 2025-10-09
+## [0.0.3] - 2025-10-10
+
+### Added
+- [CAPSEM] PII Security Policy that can block, confirm, or log based on detected PII types in model and tools responses.
+
+- [CAPSEM] config/ driven policy configuration for easy customization.
+
+- [CAPSEM Proxy] `run-proxy.py` script to run proxy in production mode.
+
+## [0.0.2] - 2025-10-09
 
 ### Added